rename manager to "buildpacks"
