+++ lib/modules/manager/api.ts
@@ -12,6 +12,7 @@ import * as bicep from './bicep';
 import * as bitbucketPipelines from './bitbucket-pipelines';
 import * as bitrise from './bitrise';
 import * as buildkite from './buildkite';
+import * as buildpacks from './buildpacks';
 import * as bun from './bun';
 import * as bunVersion from './bun-version';
 import * as bundler from './bundler';
@@ -20,7 +21,6 @@ import * as cargo from './cargo';
 import * as cdnurl from './cdnurl';
 import * as circleci from './circleci';
 import * as cloudbuild from './cloudbuild';
-import * as cnb from './cnb';
 import * as cocoapods from './cocoapods';
 import * as composer from './composer';
 import * as conan from './conan';
@@ -117,6 +117,7 @@ api.set('bicep', bicep);
 api.set('bitbucket-pipelines', bitbucketPipelines);
 api.set('bitrise', bitrise);
 api.set('buildkite', buildkite);
+api.set('buildpacks', buildpacks);
 api.set('bun', bun);
 api.set('bun-version', bunVersion);
 api.set('bundler', bundler);
@@ -125,7 +126,6 @@ api.set('cargo', cargo);
 api.set('cdnurl', cdnurl);
 api.set('circleci', circleci);
 api.set('cloudbuild', cloudbuild);
-api.set('cnb', cnb);
 api.set('cocoapods', cocoapods);
 api.set('composer', composer);
 api.set('conan', conan);
